Use non-quantized and utilization goes way up for fcn-resnet
