+++ .gitignore
@@ -290,6 +290,4 @@ __pycache__/
 
 # UWP Generated files
 **/Generated Files/
-/Samples/StyleTransfer/Assets
-/Samples/BackgroundBlur/BackgroundBlur/AsyncMFTWrapper.h
-/Samples/BackgroundBlur/BackgroundBlur/AsyncMFTWrapper.cpp
+/Samples/BackgroundBlur/BackgroundBlur/Assets
